must cast long argument to vnode_pager_setsize
SCCS-vsn: 7.74

+++ sys/ufs/lfs/lfs_vnops.c
@@ -30,7 +30,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)lfs_vnops.c	7.73 (Berkeley) 12/14/91
+ *	@(#)lfs_vnops.c	7.74 (Berkeley) 12/19/91
  */
 
 #include <sys/param.h>
@@ -330,7 +330,7 @@ lfs_write(vp, uio, ioflag, cred)
 		bn = bp->b_blkno;
 		if (uio->uio_offset + n > ip->i_size) {
 			ip->i_size = uio->uio_offset + n;
-			vnode_pager_setsize(vp, ip->i_size);
+			vnode_pager_setsize(vp, (u_long)ip->i_size);
 		}
 		size = blksize(fs);
 		(void) vnode_pager_uncache(vp);
